refactor out the main index thing
