+++ index.js
@@ -1,12 +1,10 @@
 'use strict'
 var pull = require('pull-stream')
-var query = require('./query')
-var select = require('./select')
-var mfr = require('map-filter-reduce')
 var keys = require('map-filter-reduce/keys')
 var u = require('./util')
-
-var FlumeViewLevel = require('flumeview-level')
+var Explain = require('./explain')
+var Filter = require('./filter')
+var Indexes = require('./indexes')
 
 var isArray = Array.isArray
 var isNumber = function (n) { return 'number' === typeof n }
@@ -25,10 +23,7 @@ function clone (obj) {
 module.exports = function (version, opts) {
   if(!isNumber(version)) throw new Error('flumeview-query:version expected as first arg')
   if(!isObject(opts)) throw new Error('flumeview-query: expected opts as second arg')
-  var filter = opts.filter || function () { return true }
-  var map = opts.map || function (item) { return item }
-  var exact = opts.exact !== false
-
+//  var exact = opts.exact !== false
   //answer this query by reading the entire log.
   //not efficient, but still returns the correct answer
   function fullScan (log, opts) {
@@ -56,7 +51,7 @@ module.exports = function (version, opts) {
       get: log.get,
       methods: { get: 'async', read: 'source'},
       read: function (opts) {
-        return createFilter(fullScan(log, opts), opts)
+        return Filter(fullScan(log, opts), opts)
       },
       createSink: function (cb) {return pull.onEnd(cb) }
     }
@@ -65,92 +60,20 @@ module.exports = function (version, opts) {
   return function (log, name) {
     if(!log.filename || !opts.indexes.length) return createMemoryIndex(log, name)
 
+    var view = Indexes(version, opts)(log, name)
 
-    var indexes = opts.indexes.map(function (e) {
-      return {
-        key: e.key,
-        value: e.value,
-        createStream: function (opts) {
-          opts = clone(opts)
-          opts.lte.unshift(e.key)
-          opts.gte.unshift(e.key)
-          if(!(opts.lte[0] == e.key && opts.gte[0] == e.key))
-            throw new Error('index key missing from options')
-          return read(opts)
-        }
-      }
+    view.methods.explain = 'sync'
+    view.explain = Explain(view.indexes, function (opts) {
+      opts.seqs = false; opts.values = true
+      return log.stream(opts)
     })
 
-    var view = FlumeViewLevel(version || 2, function (data, seq) {
-      data = map(data)
-      if (!filter(data)) return []
-      var A = []
-      indexes.forEach(function (index) {
-        var a = [index.key]
-        for(var i = 0; i < index.value.length; i++) {
-          var key = index.value[i]
-          if(!u.has(key, data)) return []
-          a.push(u.get(key, data))
-        }
-        if(!exact) a.push(seq);
-        A.push(a)
-      })
-      return A
-    })(log, name)
-
-
-
-    var read = view.read
-    view.methods.explain = 'sync'
-    view.explain = function (opts) {
-
-      opts = opts || {}
-      var q, k, sort
-
-      if(isArray(opts.query)) {
-        q = opts.query[0].$filter || {}
-        sort = opts.query[opts.query.length-1].$sort
-      }
-      else if(opts.query) {
-        q = opts.query
-      }
-      else
-        q = {}
-
-      var index = sort ? u.findByPath(indexes, sort) : select(indexes, q)
-
-      if(sort && !index) return pull.error(new Error('could not sort by:'+JSON.stringify(sort)))
-
-      if(!index) return {scan: true}
-      var _opts = query(index, q, exact)
-      _opts.values = true
-      _opts.keys = true
-      _opts.reverse = !!opts.reverse
-      //same default logic as pull-live
-      _opts.old = (opts.old !== false)
-      _opts.live = (opts.live === true || opts.old === false)
-      _opts.createStream = index.createStream
-      //TODO test coverage for live/old
-      _opts.sync = opts.sync
-      return _opts
+   view.read = function (opts) {
+      var _opts = view.explain(opts)
+      return Filter(_opts.createStream(_opts), opts)
     }
 
-    view.read = function (opts) {
-      var _opts = view.explain(opts = opts || {})
-      return createFilter(_opts.scan
-        ? fullScan(log, opts)
-        : pull(
-            _opts.createStream(_opts),
-            pull.map(function (data) {
-              return data.sync ? data : map(data.value)
-            })
-          ),
-        opts
-      )
-    }
     return view
   }
 }
-
-
